Corrected the Documentation Removed the mentions of private members Specifies the double is recasted to an int in the documentation of mutliplication.

+++ src/network/utils/data-rate.cc
@@ -223,7 +223,7 @@ DataRate& DataRate::operator -= (DataRate rhs)
 
 DataRate DataRate::operator * (double rhs)
 {
-  return DataRate(m_bps * rhs);
+  return DataRate(((uint64_t)(m_bps * rhs));
 }
 
 DataRate& DataRate::operator *= (double rhs)
